+++ src/main/kotlin/com/flab/inqueue/application/controller/EventController.kt
@@ -1,13 +1,9 @@
 package com.flab.inqueue.application.controller
 
-import com.flab.inqueue.domain.queue.dto.WaitQueueInfo
-import com.flab.inqueue.domain.queue.dto.JobResponse
 import com.flab.inqueue.domain.event.dto.EventRequest
 import com.flab.inqueue.domain.event.dto.EventResponse
 import com.flab.inqueue.domain.event.service.EventService
-import com.flab.inqueue.domain.queue.entity.JobStatus
 import jakarta.validation.Valid
-import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.*
 
 @RestController
@@ -23,42 +19,5 @@ class EventController(
         return eventService.save(eventRequest)
     }
 
-    @PostMapping("/{eventId}/enter")
-    fun enterWaitQueue(
-        @RequestHeader("Authorization") accessToken: String,
-        @PathVariable("eventId") eventId: String,
-    ): JobResponse {
-        return JobResponse(JobStatus.WAIT, WaitQueueInfo(1L, 1))
-    }
-
-
-    @GetMapping("/{eventId}")
-    fun retrieveWaitQueue(
-        @RequestHeader("Authorization") accessToken: String,
-        @RequestHeader("X-Client-Id") clientId: String,
-        @PathVariable eventId: String,
-    ): JobResponse {
-        return JobResponse(JobStatus.WAIT, WaitQueueInfo(1L, 1))
-    }
-
-
-    @PostMapping("/{eventId}/job-queue-check/{userId}")
-    fun validateJobQueue(
-        @RequestHeader("Authorization") accessKey: String,
-        @PathVariable eventId: String,
-        @PathVariable userId: String,
-    ): ResponseEntity<Unit> {
-        return ResponseEntity.ok().build()
-    }
-
-
-    @PostMapping("/{eventId}/job-queue-finish/{userId}")
-    fun closeJopQueue(
-        @RequestHeader("Authorization") accessKey: String,
-        @PathVariable eventId: String,
-        @PathVariable userId: String,
-    ): ResponseEntity<Unit> {
-        return ResponseEntity.ok().build()
-    }
 }
